pyproject.toml
[tool.poetry]
name = "pandoc_pdf"
version = "0.1.1"
description = "Command to generate pdf easily with pandoc."
license = "MIT"
authors = ["rai"]
packages = [
    { include = "pandoc_pdf", from = "src" },
    { include = "utils", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pandoc_pdf = 'pandoc_pdf.main:pandoc_pdf'
